run/load_to_bq/requirements.txt
Flask==2.0.1
pytest==5.3.0; python_version > "3.0"
pytest==4.6.6; python_version < "3.0"
gunicorn==20.1.0
google-cloud-storage==1.42.2
google-cloud-bigquery==2.30.1
Wand==0.6.7
